Refactor pcapng: move SimplePacket to its own file
