+++ src/pcapng.rs
@@ -51,10 +51,12 @@ use std::convert::TryFrom;
 mod enhanced_packet;
 mod interface_description;
 mod section_header;
+mod simple_packet;
 
 pub use enhanced_packet::*;
 pub use interface_description::*;
 pub use section_header::*;
+pub use simple_packet::*;
 
 trait PcapNGBlockParser<'a, En: PcapEndianness, O: 'a> {
     /// Minimum header size, in bytes
@@ -279,73 +281,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     let ts_fractional = (ts % resolution) as u32;
     // XXX should we round to closest unit?
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
-}
-
-/// The Simple Packet Block (SPB) is a lightweight container for storing
-/// the packets coming from the network.
-///
-/// This struct is a thin abstraction layer, and stores the raw block data.
-/// For ex the `data` field is stored with the padding.
-/// It implements the `PcapNGPacketBlock` trait, which provides helper functions.
-#[derive(Debug)]
-pub struct SimplePacketBlock<'a> {
-    /// Block type (little endian)
-    pub block_type: u32,
-    pub block_len1: u32,
-    /// Original packet length
-    pub origlen: u32,
-    pub data: &'a [u8],
-    pub block_len2: u32,
-}
-
-impl<'a> PcapNGPacketBlock for SimplePacketBlock<'a> {
-    fn big_endian(&self) -> bool {
-        self.block_type != SPB_MAGIC
-    }
-    fn truncated(&self) -> bool {
-        self.origlen as usize <= self.data.len()
-    }
-    fn orig_len(&self) -> u32 {
-        self.origlen
-    }
-    fn raw_packet_data(&self) -> &[u8] {
-        self.data
-    }
-    fn packet_data(&self) -> &[u8] {
-        let caplen = self.origlen as usize;
-        if caplen < self.data.len() {
-            &self.data[..caplen]
-        } else {
-            self.data
-        }
-    }
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, SimplePacketBlock<'a>>
-    for SimplePacketBlock<'a>
-{
-    const HDR_SZ: usize = 16;
-    const MAGIC: u32 = SPB_MAGIC;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], SimplePacketBlock<'a>, E> {
-        // caller function already tested header type(magic) and length
-        // read end of header
-        let (i, origlen) = En::parse_u32(i)?;
-        let (i, data) = take((block_len1 as usize) - 16)(i)?;
-        let block = SimplePacketBlock {
-            block_type,
-            block_len1,
-            origlen,
-            data,
-            block_len2,
-        };
-        Ok((i, block))
-    }
 }
 
 #[derive(Clone, Copy, Eq, PartialEq)]
@@ -873,21 +808,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-/// Parse a Simple Packet Block (little-endian)
-///
-/// *Note: this function does not remove padding in the `data` field.
-/// Use `packet_data` to get field without padding.*
-pub fn parse_simplepacketblock_le(i: &[u8]) -> IResult<&[u8], SimplePacketBlock, PcapError<&[u8]>> {
-    ng_block_parser::<SimplePacketBlock, PcapLE, _, _>()(i)
-}
-
-/// Parse a Simple Packet Block (big-endian)
-///
-/// *Note: this function does not remove padding*
-pub fn parse_simplepacketblock_be(i: &[u8]) -> IResult<&[u8], SimplePacketBlock, PcapError<&[u8]>> {
-    ng_block_parser::<SimplePacketBlock, PcapBE, _, _>()(i)
 }
 
 fn parse_name_record<'a, En: PcapEndianness, E: ParseError<&'a [u8]>>(
